[V1][Minor] Move scheduler outputs to a separate file (#13062)
